✨ feat(Dashboard): refactor Sidebar and create SidebarUserSection for improved user display and role-based navigation
